+++ code/test/sixsq/nuvla/server/resources/lifecycle_test_utils.clj
@@ -37,6 +37,7 @@
     (org.elasticsearch.common.settings Settings)
     (org.elasticsearch.index.reindex ReindexPlugin)
     (org.elasticsearch.node MockNode)
+    (org.elasticsearch.painless PainlessPlugin)
     (org.elasticsearch.transport Netty4Plugin)))
 
 
@@ -368,7 +369,8 @@
                       (put "network.host" "127.0.0.1")
                       (build))
          plugins  [Netty4Plugin
-                   ReindexPlugin]]
+                   ReindexPlugin
+                   PainlessPlugin]]
 
      (LogConfigurator/configureWithoutConfig settings)
      (.. (MockNode. ^Settings settings plugins)