+++ code/test/sixsq/nuvla/server/resources/module_applications_sets_lifecycle_test.clj
@@ -1,4 +1,4 @@
-(ns sixsq.nuvla.server.resources.module-application-lifecycle-test
+(ns sixsq.nuvla.server.resources.module-applications-sets-lifecycle-test
   (:require
     [clojure.data.json :as json]
     [clojure.test :refer [deftest use-fixtures]]
@@ -7,13 +7,13 @@
     [sixsq.nuvla.server.middleware.authn-info :refer [authn-info-header]]
     [sixsq.nuvla.server.resources.common.utils :as u]
     [sixsq.nuvla.server.resources.lifecycle-test-utils :as ltu]
-    [sixsq.nuvla.server.resources.module-application :as module-application]))
+    [sixsq.nuvla.server.resources.module-applications-sets :as t]))
 
 
 (use-fixtures :once ltu/with-test-server-fixture)
 
 
-(def base-uri (str p/service-context module-application/resource-type))
+(def base-uri (str p/service-context t/resource-type))
 
 
 (def valid-acl {:owners ["group/nuvla-admin"]})
@@ -22,18 +22,18 @@
 (def timestamp "1964-08-25T10:00:00.00Z")
 
 
-(def valid-entry {:id                  (str module-application/resource-type "/module-application-uuid")
-                  :resource-type       module-application/resource-type
-                  :created             timestamp
-                  :updated             timestamp
-                  :acl                 valid-acl
+(def valid-entry {:id                (str t/resource-type "/module-applications-sets-uuid")
+                  :resource-type     t/resource-type
+                  :created           timestamp
+                  :updated           timestamp
+                  :acl               valid-acl
 
-                  :author              "someone"
-                  :commit              "wip"
+                  :author            "someone"
+                  :commit            "wip"
 
-                  :unsupported-options ["devices"]
-
-                  :docker-compose      "version: \"3.3\"\nservices:\n  web:\n    ..."})
+                  :applications-sets [{:name         "x"
+                                       :applications [{:id      "module/x"
+                                                       :version 0}]}]})
 
 
 (deftest lifecycle
@@ -112,6 +112,6 @@
 
 
 (deftest bad-methods
-  (let [resource-uri (str p/service-context (u/new-resource-id module-application/resource-type))]
+  (let [resource-uri (str p/service-context (u/new-resource-id t/resource-type))]
     (ltu/verify-405-status [[base-uri :delete]
                             [resource-uri :post]])))