+++ code/test/sixsq/nuvla/server/resources/module_lifecycle_test.clj
@@ -10,32 +10,11 @@
     [sixsq.nuvla.server.resources.module :as module]
     [sixsq.nuvla.server.resources.module.utils :as utils]))
 
-
-(use-fixtures :once ltu/with-test-server-fixture)
-
+(use-fixtures :each ltu/with-test-server-fixture)
 
 (def base-uri (str p/service-context module/resource-type))
 
-
 (def timestamp "1964-08-25T10:00:00.00Z")
-
-(defn valid-module
-  [subtype content]
-  {:id                        (str module/resource-type "/connector-uuid")
-   :resource-type             module/resource-type
-   :created                   timestamp
-   :updated                   timestamp
-   :parent-path               "a/b"
-   :path                      "a/b/c"
-   :subtype                   subtype
-
-   :logo-url                  "https://example.org/logo"
-
-   :data-accept-content-types ["application/json" "application/x-something"]
-   :data-access-protocols     ["http+s3" "posix+nfs"]
-
-   :content                   content})
-
 
 (defn lifecycle-test-module
   [subtype valid-content]
@@ -46,13 +25,11 @@
         session-user  (header session-anon authn-info-header
                               "user/jane user/jane group/nuvla-user group/nuvla-anon")
 
-        valid-entry   {:id                        (str module/resource-type "/connector-uuid")
-                       :resource-type             module/resource-type
-                       :created                   timestamp
-                       :updated                   timestamp
-                       :parent-path               "a/b"
+        valid-entry   {:parent-path               "a/b"
                        :path                      "a/b/c"
                        :subtype                   subtype
+
+                       :compatibility             "docker-compose"
 
                        :logo-url                  "https://example.org/logo"
 
@@ -90,6 +67,16 @@
         (ltu/body->edn)
         (ltu/is-status 400))
 
+    (when (utils/is-application? subtype)
+      (testing "application should have compatibility attribute set"
+        (-> session-user
+            (request base-uri
+                     :request-method :post
+                     :body (json/write-str (dissoc valid-entry :compatibility)))
+            (ltu/body->edn)
+            (ltu/is-status 400)
+            (ltu/message-matches "Application subtype should have compatibility attribute set!"))))
+
     ;; adding, retrieving and  deleting entry as user should succeed
     (doseq [session [session-admin session-user]]
       (let [uri     (-> session
@@ -112,13 +99,10 @@
                           (request abs-uri)
                           (ltu/body->edn)
                           (ltu/is-status 200)
+                          (ltu/is-key-value :compatibility "docker-compose")
                           (as-> m (if (utils/is-application? subtype)
-                                    (do
-                                      (ltu/is-key-value m :compatibility "docker-compose")
-                                      (ltu/is-operation-present m :validate-docker-compose))
-                                    (do
-                                      (ltu/is-key-value m :compatibility nil)
-                                      (ltu/is-operation-absent m :validate-docker-compose))))
+                                    (ltu/is-operation-present m :validate-docker-compose)
+                                    (ltu/is-operation-absent m :validate-docker-compose)))
                           (ltu/body)
                           :content)]
           (is (= valid-content (select-keys content (keys valid-content)))))
@@ -253,16 +237,16 @@
               (ltu/is-key-value :published false)
               (ltu/get-op-url :unpublish)))
 
-        ;; edit module without putting the module-content should not create new version
-        (is (= 7 (-> session-admin
-                     (request abs-uri
-                              :request-method :put
-                              :body (json/write-str (dissoc valid-entry :content :path)))
-                     (ltu/body->edn)
-                     (ltu/is-status 200)
-                     (ltu/body)
-                     :versions
-                     count)))
+        (testing "edit module without putting the module-content should not create new version"
+          (is (= 7 (-> session-admin
+                       (request abs-uri
+                                :request-method :put
+                                :body (json/write-str (dissoc valid-entry :content :path)))
+                       (ltu/body->edn)
+                       (ltu/is-status 200)
+                       (ltu/body)
+                       :versions
+                       count))))
 
         (doseq [i ["_0/delete-version" "_1/delete-version"]]
           (-> session-admin
@@ -278,15 +262,15 @@
 
         (testing "delete latest version without specifying version"
           (-> session-admin
-             (request (str abs-uri "/delete-version"))
-             (ltu/body->edn)
-             (ltu/is-status 200)))
+              (request (str abs-uri "/delete-version"))
+              (ltu/body->edn)
+              (ltu/is-status 200)))
 
         (testing "delete out of bound index should return 404"
           (-> session-admin
-             (request (str abs-uri "_50/delete-version"))
-             (ltu/body->edn)
-             (ltu/is-status 404)))
+              (request (str abs-uri "_50/delete-version"))
+              (ltu/body->edn)
+              (ltu/is-status 404)))
 
         (-> session-admin
             (request (str abs-uri "_50"))
@@ -329,9 +313,88 @@
 
   (let [valid-application {:author         "someone"
                            :commit         "wip"
-
                            :docker-compose "version: \"3.6\"\n\nx-common: &common\n  stop_grace_period: 4s\n  logging:\n    options:\n      max-size: \"250k\"\n      max-file: \"10\"\n  labels:\n    - \"nuvlabox.component=True\"\n    - \"nuvlabox.deployment=production\"\n\nvolumes:\n  nuvlabox-db:\n    driver: local\n\nnetworks:\n  nuvlabox-shared-network:\n    driver: overlay\n    name: nuvlabox-shared-network\n    attachable: true\n\nservices:\n  data-gateway:\n    <<: *common\n    image: traefik:2.1.1\n    container_name: datagateway\n    restart: on-failure\n    command:\n      - --entrypoints.mqtt.address=:1883\n      - --entrypoints.web.address=:80\n      - --providers.docker=true\n      - --providers.docker.exposedbydefault=false\n    volumes:\n      - /var/run/docker.sock:/var/run/docker.sock\n    networks:\n      - default\n      - nuvlabox-shared-network\n\n  nb-mosquitto:\n    <<: *common\n    image: eclipse-mosquitto:1.6.8\n    container_name: nbmosquitto\n    restart: on-failure\n    labels:\n      - \"traefik.enable=true\"\n      - \"traefik.tcp.routers.mytcprouter.rule=HostSNI(`*`)\"\n      - \"traefik.tcp.routers.mytcprouter.entrypoints=mqtt\"\n      - \"traefik.tcp.routers.mytcprouter.service=mosquitto\"\n      - \"traefik.tcp.services.mosquitto.loadbalancer.server.port=1883\"\n      - \"nuvlabox.component=True\"\n      - \"nuvlabox.deployment=production\"\n    healthcheck:\n      test: [\"CMD-SHELL\", \"timeout -t 5 mosquitto_sub -t '$$SYS/#' -C 1 | grep -v Error || exit 1\"]\n      interval: 10s\n      timeout: 10s\n      start_period: 10s\n\n  system-manager:\n    <<: *common\n    image: nuvlabox/system-manager:1.0.1\n    restart: always\n    volumes:\n      - /var/run/docker.sock:/var/run/docker.sock\n      - nuvlabox-db:/srv/nuvlabox/shared\n    ports:\n      - 127.0.0.1:3636:3636\n    healthcheck:\n      test: [\"CMD\", \"curl\", \"-f\", \"http://localhost:3636\"]\n      interval: 30s\n      timeout: 10s\n      retries: 4\n      start_period: 10s\n\n  agent:\n    <<: *common\n    image: nuvlabox/agent:1.3.2\n    restart: on-failure\n    environment:\n      - NUVLABOX_UUID=${NUVLABOX_UUID}\n      - NUVLA_ENDPOINT=${NUVLA_ENDPOINT:-nuvla.io}\n      - NUVLA_ENDPOINT_INSECURE=${NUVLA_ENDPOINT_INSECURE:-False}\n    volumes:\n      - /var/run/docker.sock:/var/run/docker.sock\n      - nuvlabox-db:/srv/nuvlabox/shared\n      - /:/rootfs:ro\n    expose:\n      - 5000\n    depends_on:\n      - system-manager\n      - compute-api\n\n  management-api:\n    <<: *common\n    image: nuvlabox/management-api:0.1.0\n    restart: on-failure\n    environment:\n      - NUVLA_ENDPOINT=${NUVLA_ENDPOINT:-nuvla.io}\n      - NUVLA_ENDPOINT_INSECURE=${NUVLA_ENDPOINT_INSECURE:-False}\n    volumes:\n      - /proc/sysrq-trigger:/sysrq\n      - ${HOME}/.ssh/authorized_keys:/rootfs/.ssh/authorized_keys\n      - nuvlabox-db:/srv/nuvlabox/shared\n      - /var/run/docker.sock:/var/run/docker.sock\n    ports:\n      - 5001:5001\n    healthcheck:\n      test: curl -k https://localhost:5001 2>&1 | grep SSL\n      interval: 20s\n      timeout: 10s\n      start_period: 30s\n\n  compute-api:\n    <<: *common\n    image: nuvlabox/compute-api:0.2.5\n    restart: on-failure\n    pid: \"host\"\n    environment:\n      - HOST=${HOSTNAME:-nuvlabox}\n    volumes:\n      - /var/run/docker.sock:/var/run/docker.sock\n      - nuvlabox-db:/srv/nuvlabox/shared\n    ports:\n      - 5000:5000\n    depends_on:\n      - system-manager\n\n  network-manager:\n    <<: *common\n    image: nuvlabox/network-manager:0.0.4\n    restart: on-failure\n    environment:\n      - NUVLABOX_UUID=${NUVLABOX_UUID}\n      - VPN_INTERFACE_NAME=${NUVLABOX_VPN_IFACE:-vpn}\n    volumes:\n      - nuvlabox-db:/srv/nuvlabox/shared\n    depends_on:\n      - system-manager\n\n  vpn-client:\n    <<: *common\n    image: nuvlabox/vpn-client:0.0.4\n    container_name: vpn-client\n    restart: always\n    network_mode: host\n    cap_add:\n      - NET_ADMIN\n    devices:\n      - /dev/net/tun\n    environment:\n      - NUVLABOX_UUID=${NUVLABOX_UUID}\n    volumes:\n      - nuvlabox-db:/srv/nuvlabox/shared\n    depends_on:\n      - network-manager"}]
     (lifecycle-test-module utils/subtype-app valid-application)))
+
+
+(def valid-applications-sets-content
+  {:author            "someone"
+   :commit            "wip"
+   :applications-sets [{:name         "x"
+                        :applications [{:id      "module/x"
+                                        :version 0}]}]})
+
+(deftest lifecycle-applications-sets
+  (lifecycle-test-module utils/subtype-apps-sets valid-applications-sets-content))
+
+
+(deftest lifecycle-applications-sets-extended
+  (let [session-anon      (-> (session (ltu/ring-app))
+                              (content-type "application/json"))
+        session-user      (header session-anon authn-info-header
+                                  "user/jane user/jane group/nuvla-user group/nuvla-anon")
+
+        valid-app-1       {:parent-path   "a/b"
+                           :path          "clara/app-1"
+                           :subtype       utils/subtype-app
+                           :compatibility "docker-compose"
+                           :content       {:author         "someone"
+                                           :commit         "initial"
+                                           :docker-compose "some content"}}
+        app-1-create-resp (-> session-user
+                              (request base-uri
+                                       :request-method :post
+                                       :body (json/write-str valid-app-1))
+                              (ltu/body->edn)
+                              (ltu/is-status 201))
+        app-1-uri         (ltu/location-url app-1-create-resp)
+        app-1-id          (ltu/location app-1-create-resp)]
+
+    (let [valid-entry {:parent-path "a/b"
+                       :path        "a/b/c"
+                       :subtype     utils/subtype-apps-sets
+                       :content     (assoc-in valid-applications-sets-content
+                                              [:applications-sets 0
+                                               :applications 0 :id] app-1-id)}]
+
+      (-> session-user
+          (request app-1-uri
+                   :request-method :put
+                   :body (json/write-str
+                           (update valid-app-1 :content assoc
+                                   :docker-compose "content changed"
+                                   :commit "second commit")))
+          (ltu/body->edn)
+          (ltu/is-status 200))
+
+      (let [response   (-> session-user
+                           (request base-uri
+                                    :request-method :post
+                                    :body (json/write-str valid-entry))
+                           (ltu/body->edn)
+                           (ltu/is-status 201))
+            uri        (ltu/location response)
+            abs-uri    (ltu/location-url response)
+            deploy-uri (-> session-user
+                           (request abs-uri)
+                           (ltu/body->edn)
+                           (ltu/is-status 200)
+                           (ltu/get-op-url :deploy))]
+        (-> session-user
+            (request deploy-uri)
+            (ltu/body->edn)
+            (ltu/is-status 200)
+            (ltu/is-key-value :application uri)
+            (ltu/is-key-value :version 0)
+            (ltu/is-key-value #(-> %
+                                   first
+                                   :applications
+                                   first
+                                   :resolved
+                                   :content
+                                   :docker-compose)
+                              :applications-sets
+                              "some content"))))))
 
 
 (deftest bad-methods