+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 ## Unreleased
 
+### Changed
+
+## [6.0.12] - 2023-04-24
+
+### Changed
+
+- Module - Let user set compatibility flag
+- Module application - Deprecate unsupported-options
+- ES binding - Support bulk-edit
+- Spec helper - Generic spec validator
+- Spec helper - Request body validator
+- Common body - Spec for common request body specs
+- Nuvlabox - Adds bulk editing endpoints for tags
+- Module - Resolve vendor email
+- Module - New subtype applications set
+- Deployment set - Experimental v2
+
 ## [6.0.11] - 2023-02-22
 
 ### Changed